refactor: organize property reports
